requirement.txt
scikit-learn==0.20.0
tensorflow-gpu==2.0.0
tensorflow-addons==0.6.0; sys_platform == 'linux'
numpy==1.16.2
matplotlib==3.0.0
tinydb==3.11.1
xlrd==1.2.0
xlwt==1.3.0
pandas==0.24.2
PILLOW==6.0.0
tqdm==4.32.1
fpdf==1.7.2
seaborn==0.9.0
opencv-python>=4.2
pyper==1.1.2
more_itertools
